update Layout file structure
