+++ frontend/src/components/Layout/Header/Header.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
 
-const Header = () => {
+const Header: React.FC = () => {
   const location = useLocation()
   return (
     <div id="header">
