+++ frontend/src/components/Layout/Header/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Header'
