+++ frontend/src/components/Layout/Modal/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Modal'
