+++ frontend/src/components/Layout/Navbar/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Navbar'
