+++ frontend/src/components/Layout/Notification/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Notification'
